1.12

+++ heegaard_floer.h
@@ -11,6 +11,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
+#include <climits>
 
 #include "grid_diagram.h"
 #include "grid_state.h"
@@ -32,11 +33,23 @@ vector<GridState> GetAllGridState (const GridDiagram& grid_diagram);
 // @param: grid_diagram A grid diagram;
 void PrintAllGridState (const GridDiagram& grid_diagram);   
 
-int GetXMaslovGrading (const GridDiagram& grid_diagram, const GridState& grid_state);
+// @brief: Get Maslov X Grading of a GridState;
+// @param: grid_diagram A grid diagram;
+// @param: grid_state A grid state inside this diagram;
+// @return: Maslov X grading (half integer, double); 
+double GetXMaslovGrading (const GridDiagram& grid_diagram, const GridState& grid_state);
 
-int GetOMaslovGrading (const GridDiagram& grid_diagram, const GridState& grid_state);  
-  
-int GetReducedAlexanderGrading (const GridDiagram& grid_diagram, const GridState& grid_state);
+// @brief: Get Maslov X Grading of a GridState;
+// @param: grid_diagram A grid diagram;
+// @param: grid_state A grid state inside this diagram;
+// @return: Maslov X grading (half integer, double);   
+double GetOMaslovGrading (const GridDiagram& grid_diagram, const GridState& grid_state);  
+
+// @brief: Get Alexander Grading of a GridState;
+// @param: grid_diagram A grid diagram;
+// @param: grid_state A grid state inside this diagram;
+// @return: Alexander grading (half integer, double);   
+double GetReducedAlexanderGrading (const GridDiagram& grid_diagram, const GridState& grid_state);
   
 }  // namespace heegaardfloer;
 }  // namespace gridhomology;
